Move function for generating mocked cost data to `testUtils.ts`
Acked-by: Akira Hayashi <rindrics@gmail.com>
